Functions that have been renamed, superseded, or are no longer working

Generally these functions have been superseded or renamed. Upon using them, a message is provided directing the user to the new function. However, at this stage of package development, we generally clear older defunct functions at each minor release, and so you are strongly encouraged to use the new functions/names/syntax wherever possible and update your scripts accordingly.

Usage

coalesce_compatible(.data)

network_map(.data)

db_plot(.data)

db_comp(.data)

retrieve_bilaterals(.data)

retrieve_multilaterals(.data)

retrieve_links(.data)

retrieve_membership_list(.data)

data_source(.data)

get_packages(.data)

open_codebook(.data)

data_contrast(.data)

data_evolution(.data)

plot_releases(.data)

Functions

coalesce_compatible(): Deprecated on 2022-10-18.

network_map(): Removed on 2023-06-15.

db_plot(): Removed on 2023-06-19.

db_comp(): Removed on 2023-06-19.

retrieve_bilaterals(): Removed on 2023-06-19.

retrieve_multilaterals(): Removed on 2023-06-19.

retrieve_links(): Removed on 2023-06-19.

retrieve_membership_list(): Removed on 2023-06-19.

data_source(): Removed on 2023-06-19.

get_packages(): Removed on 2023-06-19.

open_codebook(): Removed on 2023-06-19.

data_contrast(): Removed on 2023-06-19.

data_evolution(): Removed on 2023-06-19.

plot_releases(): Renamed on 2023-07-03.