Emperors datacube documentation

Emperors datacube documentation

Usage

emperors

Format

The emperors datacube is a list that contains the following 3 datasets: wikipedia, UNRV, britannica. For more information and references to each of the datasets used, please use the data_source() and data_contrast(). functions.

wikipedia:

A dataset with 68 observations and the following 15 variables: ID, Begin, End, FullName, Birth, Death, CityBirth, ProvinceBirth, Rise, Cause, Killer, Dynasty, Era, Notes, Verif.

UNRV:

A dataset with 99 observations and the following 7 variables: ID, Begin, End, Birth, Death, FullName, Dynasty.

britannica:

A dataset with 87 observations and the following 3 variables: ID, Begin, End.

Source

wikipedia: Wikipedia, List_of_Roman_emperors, https://en.wikipedia.org/wiki/List_of_Roman_emperors, Accessed on 2021-07-22

UNRV: UNRV, Roman Emperor list, https://www.unrv.com/government/emperor.php, Accessed on 2021-07-22

britannica: Britannica, List of Roman emperors, https://www.britannica.com/topic/list-of-Roman-emperors-2043294, Accessed on 2021-07-22

Details

#> $wikipedia
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |      ID       |  character  |              68|     0     |       0       |
#> |     Begin     |    mdate    |              68|     0     |       0       |
#> |      End      |    mdate    |              68|     0     |       0       |
#> |   FullName    |  character  |              68|     0     |       0       |
#> |     Birth     |  character  |              68|     5     |     7.35      |
#> |     Death     |  character  |              68|     0     |       0       |
#> |   CityBirth   |  character  |              68|    17     |      25       |
#> | ProvinceBirth |  character  |              68|     0     |       0       |
#> |     Rise      |  character  |              68|     0     |       0       |
#> |     Cause     |  character  |              68|     0     |       0       |
#> |    Killer     |  character  |              68|     0     |       0       |
#> |    Dynasty    |  character  |              68|     0     |       0       |
#> |      Era      |  character  |              68|     0     |       0       |
#> |     Notes     |  character  |              68|    22     |     32.35     |
#> |     Verif     |  character  |              68|    57     |     83.82     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $UNRV
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |      ID       |  character  |              99|     0     |       0       |
#> |     Begin     |    mdate    |              99|     0     |       0       |
#> |      End      |    mdate    |              99|     0     |       0       |
#> |     Birth     |  character  |              99|     0     |       0       |
#> |     Death     |  character  |              99|     0     |       0       |
#> |   FullName    |  character  |              99|     5     |     5.05      |
#> |    Dynasty    |  character  |              99|    37     |     37.37     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $britannica
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |      ID       |  character  |              87|     0     |       0       |
#> |     Begin     |    mdate    |              87|     0     |       0       |
#> |      End      |    mdate    |              87|     0     |       0       |
#> ----------------------------------------------------------------------------

URL

wikipedia: https://en.wikipedia.org/wiki/List_of_Roman_emperors

UNRV: https://www.unrv.com/government/emperor.php

britannica: https://www.britannica.com/topic/list-of-Roman-emperors-2043294

Mapping

wikipedia: Variable Mapping

from	to

name	ID

reign.start	Begin

reign.end	End

name.full	FullName

birth	Birth

death	Death

birth.cty	CityBirth

birth.prv	ProvinceBirth

rise	Rise

cause	Cause

killer	Killer

dynasty	Dynasty

era	Era

notes	Notes

verif.who	Verif

UNRV: Variable Mapping

from	to

'Common Name'	ID

Beg	Begin

'Full Name/Imperial Name'	FullName

'Dynasty/Class/Notes'	Dynasty

britannica: Variable Mapping

from	to

Name	ID

reign_start	Begin

reign_end	End