Get international treaties

Some datasets in the membership databases across the 'many* packages' (e.g. manyenviron) contain a myriad of information on international treaties governing an international domain. The get treaty functions help researchers retrieve all bilateral agreements, or multilateral agreements, from these datasets. Researchers can, for example, use get_bilaterals() to retrieve which countries have signed a specific international agreement, or several international agreements signed in a respective year. As well, researchers can use get_multilaterals() to retrieve the titles of all multilateral agreements signed in the past 10 years, for instance. Alternatively, to get information from several datasets in a memberships database, researchers can consolidate() the database into one dataset with some combination of the rows, columns, and observations of the datasets before getting the desired bilateral or multilateral treaties.

Usage

get_bilaterals(membs)

get_multilaterals(membs)

Arguments

membs

A memberships dataset from one of the many packages

Value

A tibble of bilateral agreements A tibble of multilateral agreements

Examples

membs <- tibble::tibble(CountryID = c("ROU", "RUS", "DNK"),
manyID = c("ROU-RUS[RFP]_1901A", "ROU-RUS[RFP]_1901A", "GD16FI_1901A"),
Title = c("Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between The Governments Of Denmark And
The United Kingdom Of Great Britain
And Northern Ireland For Regulating The Fisheries
Of Their Respective Subjects Outside
Territorial Waters In The Ocean Surrounding The Faroe Islands"),
Beg = c("1901-02-22", "1901-02-22", "1901-06-24"),
End = c(NA, NA, NA))
get_bilaterals(membs)
#> # A tibble: 1 × 5
#>   CountryID1 CountryID2 Title                                        Beg   End  
#>   <chr>      <chr>      <chr>                                        <chr> <lgl>
#> 1 RUS        ROU        "Convention Between Roumania And Russia Con… 1901… NA   
membs <- tibble::tibble(CountryID = c("ROU", "RUS", "DNK"),
manyID = c("ROU-RUS[RFP]_1901A", "ROU-RUS[RFP]_1901A", "GD16FI_1901A"),
Title = c("Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between The Governments Of Denmark And
The United Kingdom Of Great Britain
And Northern Ireland For Regulating The Fisheries
Of Their Respective Subjects Outside
Territorial Waters In The Ocean Surrounding The Faroe Islands"),
Beg = c("1901-02-22", "1901-02-22", "1901-06-24"),
End = c(NA, NA, NA))
get_multilaterals(membs)
#> # A tibble: 1 × 4
#>   manyID       Title                                                 Beg   End  
#>   <chr>        <chr>                                                 <chr> <lgl>
#> 1 GD16FI_1901A "Convention Between The Governments Of Denmark And\n… 1901… NA