Drop only columns used in formula

A function between dplyr's transmute and mutate

transmutate(.data, ...)

Arguments

.data
Data frame to pass to the function

...
Variables to pass to the function

Source

https://stackoverflow.com/questions/51428156/dplyr-mutate-transmute-drop-only-the-columns-used-in-the-formula

Value

Data frame with mutated variables and none of the variables used in the mutations, but, unlike dplyr::transmute(), all other unnamed variables.

Examples

transmutate(mtcars, X = ifelse( vs, drat, wt ), Y = mpg*cyl)
#>                      disp  hp  qsec am gear carb     X     Y
#> Mazda RX4           160.0 110 16.46  1    4    4 2.620 126.0
#> Mazda RX4 Wag       160.0 110 17.02  1    4    4 2.875 126.0
#> Datsun 710          108.0  93 18.61  1    4    1 3.850  91.2
#> Hornet 4 Drive      258.0 110 19.44  0    3    1 3.080 128.4
#> Hornet Sportabout   360.0 175 17.02  0    3    2 3.440 149.6
#> Valiant             225.0 105 20.22  0    3    1 2.760 108.6
#> Duster 360          360.0 245 15.84  0    3    4 3.570 114.4
#> Merc 240D           146.7  62 20.00  0    4    2 3.690  97.6
#> Merc 230            140.8  95 22.90  0    4    2 3.920  91.2
#> Merc 280            167.6 123 18.30  0    4    4 3.920 115.2
#> Merc 280C           167.6 123 18.90  0    4    4 3.920 106.8
#> Merc 450SE          275.8 180 17.40  0    3    3 4.070 131.2
#> Merc 450SL          275.8 180 17.60  0    3    3 3.730 138.4
#> Merc 450SLC         275.8 180 18.00  0    3    3 3.780 121.6
#> Cadillac Fleetwood  472.0 205 17.98  0    3    4 5.250  83.2
#> Lincoln Continental 460.0 215 17.82  0    3    4 5.424  83.2
#> Chrysler Imperial   440.0 230 17.42  0    3    4 5.345 117.6
#> Fiat 128             78.7  66 19.47  1    4    1 4.080 129.6
#> Honda Civic          75.7  52 18.52  1    4    2 4.930 121.6
#> Toyota Corolla       71.1  65 19.90  1    4    1 4.220 135.6
#> Toyota Corona       120.1  97 20.01  0    3    1 3.700  86.0
#> Dodge Challenger    318.0 150 16.87  0    3    2 3.520 124.0
#> AMC Javelin         304.0 150 17.30  0    3    2 3.435 121.6
#> Camaro Z28          350.0 245 15.41  0    3    4 3.840 106.4
#> Pontiac Firebird    400.0 175 17.05  0    3    2 3.845 153.6
#> Fiat X1-9            79.0  66 18.90  1    4    1 4.080 109.2
#> Porsche 914-2       120.3  91 16.70  1    5    2 2.140 104.0
#> Lotus Europa         95.1 113 16.90  1    5    2 3.770 121.6
#> Ford Pantera L      351.0 264 14.50  1    5    4 3.170 126.4
#> Ferrari Dino        145.0 175 15.50  1    5    6 2.770 118.2
#> Maserati Bora       301.0 335 14.60  1    5    8 3.570 120.0
#> Volvo 142E          121.0 109 18.60  1    4    2 4.110  85.6