Consolidate database into a single dataset

This function collapses a set or database of many datasets into a single dataset with some combination of the rows, columns, and observations of the parent datasets. The function includes separate arguments for the rows and columns, as well as for how to resolve conflicts in observations across datasets. This provides users with considerable flexibility in how they combine manydata. For example, users may wish to stick to units that appear in every dataset but include variables coded in any dataset, or units that appear in any dataset but only those variables that appear in every dataset. Even then there may be conflicts, as the actual unit-variable observations may differ from dataset to dataset. Here we (will) offer a number of resolve methods that enable users to choose how conflicts between observations are resolved.

consolidate(
  database,
  rows = c("any", "every"),
  cols = c("any", "every"),
  resolve = c("coalesce", "min", "max", "median", "mean", "random"),
  key = "many_ID"
)

Arguments

database
A database object from one of the many packages

rows
Which rows or units to retain. By default "any" (or all) units are retained, but another option is "every", which retains only those units that appear in all parent datasets.

cols
Which columns or variables to retain. By default "any" (or all) variables are retained, but another option is "every", which retains only those variables that appear in all parent datasets.

resolve
How should conflicts between observations be resolved? Currently "coalesce", "min", "max", "mean", "median", and "random" are offered. "coalesce" takes the first non-NA value. "max" takes the largest value, while "min" takes the smallest value. "mean" takes the average value. "median" takes the median value. "random" takes a random value. For different variables to be resolved differently, you can specify the variables' names alongside how each is to be resolved in a list. In this case, only the variables named will be resolved and returned.

key
An ID column to collapse by. By default "many_ID".

Value

A single tibble/data frame.

See also

Examples

# \donttest{
pluck(emperors, "UNRV")
#> # A tibble: 99 × 7
#>    ID               Beg       End       Birth Death FullName          Dynasty   
#>    <chr>            <messydt> <messydt> <chr> <chr> <chr>             <chr>     
#>  1 Augustus         -XX27     -XX14     63 BC 14    Gaius Julius Cae… Julio-Cla…
#>  2 Tiberius         -XX14     0037      42 BC 37    Tiberius Claudiu… Julio-Cla…
#>  3 Gaius (Caligula) 0037      0041      12    14    Gaius Caesar Ger… Julio-Cla…
#>  4 Claudius         0041      0054      10 BC 41    Tiberius Claudiu… Julio-Cla…
#>  5 Nero             0054      0068      37    68    Claudius Nero Ca… Julio-Cla…
#>  6 Galba            0068      0069      3 BC  69    Servius Sulpiciu… Year of t…
#>  7 Otho             0069      0069      32    69    Marcus Salvius O… Year of t…
#>  8 Vitellius        0069      0069      15    69    Aulus Vitellius … Year of t…
#>  9 Vespasian        0069      0079      9     79    Titus Flavius Ve… Year of t…
#> 10 Titus            0079      0081      39    79    Titus Flavius Ve… Flavian   
#> # … with 89 more rows
consolidate(emperors, "any", "any", resolve = "coalesce", key = "ID")
#> # A tibble: 137 × 15
#>    ID        CityBirth  ProvinceBirth Rise  Cause Killer Era   Notes Verif Birth
#>    <chr>     <chr>      <chr>         <chr> <chr> <chr>  <chr> <chr> <chr> <chr>
#>  1 Augustus  Rome       Italia        Birt… Assa… Wife   Prin… birt… Redd… 0062…
#>  2 Tiberius  Rome       Italia        Birt… Assa… Other… Prin… birt… Redd… 0041…
#>  3 Caligula  Antitum    Italia        Birt… Assa… Senate Prin… assa… Redd… 0012…
#>  4 Claudius  Lugdunum   Gallia Lugdu… Birt… Assa… Wife   Prin… birt… Redd… 0009…
#>  5 Nero      Antitum    Italia        Birt… Suic… Senate Prin… NA    Redd… 0037…
#>  6 Galba     Terracina  Italia        Seiz… Assa… Other… Prin… birt… Redd… 0002…
#>  7 Otho      Terentinum Italia        Appo… Suic… Other… Prin… NA    NA    0032…
#>  8 Vitellius Rome       Italia        Seiz… Assa… Other… Prin… NA    NA    0015…
#>  9 Vespasian Falacrine  Italia        Seiz… Natu… Disea… Prin… NA    NA    0009…
#> 10 Titus     Rome       Italia        Birt… Natu… Disea… Prin… NA    NA    0039…
#> # … with 127 more rows, and 5 more variables: Death <chr>, FullName <chr>,
#> #   Dynasty <chr>, Beg <messydt>, End <messydt>
consolidate(emperors, "every", "every", resolve = "min", key = "ID")
#> # A tibble: 41 × 3
#>    ID        Beg         End       
#>    <chr>     <chr>       <chr>     
#>  1 Augustus  -XX26-01-16 -XX14     
#>  2 Tiberius  -XX14       0037-01-01
#>  3 Claudius  0041-01-01  0054-01-01
#>  4 Nero      0054-01-01  0068-01-01
#>  5 Galba     0068-01-01  0069-01-01
#>  6 Otho      0069-01-01  0069-01-01
#>  7 Vespasian 0069-01-01  0079-01-01
#>  8 Titus     0079-01-01  0081-01-01
#>  9 Domitian  0081-01-01  0096-01-01
#> 10 Nerva     0096-01-01  0098-01-01
#> # … with 31 more rows
consolidate(emperors, "any", "every", resolve = "max", key = "ID")
#> # A tibble: 137 × 3
#>    ID        Beg        End       
#>    <chr>     <chr>      <chr>     
#>  1 Augustus  -XX31      0014-12-31
#>  2 Tiberius  0014-12-31 0037-12-31
#>  3 Caligula  0037-12-31 0041-12-31
#>  4 Claudius  0041-12-31 0054-12-31
#>  5 Nero      0054-12-31 0068-12-31
#>  6 Galba     0068-12-31 0069-12-31
#>  7 Otho      0069-12-31 0069-12-31
#>  8 Vitellius 0069-12-31 0069-12-31
#>  9 Vespasian 0069-12-31 0079-12-31
#> 10 Titus     0079-12-31 0081-12-31
#> # … with 127 more rows
consolidate(emperors, "every", "any", resolve = "median", key = "ID")
#> # A tibble: 41 × 15
#>    ID        CityBirth  ProvinceBirth Rise  Cause Killer Era   Notes Verif Birth
#>    <chr>     <chr>      <chr>         <chr> <chr> <chr>  <chr> <chr> <chr> <chr>
#>  1 Augustus  Rome       Italia        Birt… Assa… Wife   Prin… birt… Redd… 0062…
#>  2 Tiberius  Rome       Italia        Birt… Assa… Other… Prin… birt… Redd… 0041…
#>  3 Claudius  Lugdunum   Gallia Lugdu… Birt… Assa… Wife   Prin… birt… Redd… 0009…
#>  4 Nero      Antitum    Italia        Birt… Suic… Senate Prin… NA    Redd… 0037…
#>  5 Galba     Terracina  Italia        Seiz… Assa… Other… Prin… birt… Redd… 0002…
#>  6 Otho      Terentinum Italia        Appo… Suic… Other… Prin… NA    NA    0032…
#>  7 Vespasian Falacrine  Italia        Seiz… Natu… Disea… Prin… NA    NA    0009…
#>  8 Titus     Rome       Italia        Birt… Natu… Disea… Prin… NA    NA    0039…
#>  9 Domitian  Rome       Italia        Birt… Assa… Court… Prin… NA    NA    0051…
#> 10 Nerva     Narni      Italia        Appo… Natu… Disea… Prin… NA    NA    0030…
#> # … with 31 more rows, and 5 more variables: Death <chr>, FullName <chr>,
#> #   Dynasty <chr>, Beg <chr>, End <chr>
consolidate(emperors, "every", "every", resolve = "mean", key = "ID")
#> # A tibble: 41 × 3
#>    ID        Beg         End       
#>    <chr>     <chr>       <chr>     
#>  1 Augustus  -XX26-01-16 0014-08-19
#>  2 Tiberius  0014-09-18  0037-03-16
#>  3 Claudius  0041-01-25  0054-10-13
#>  4 Nero      0054-10-13  0068-06-09
#>  5 Galba     0068-06-08  0069-01-15
#>  6 Otho      0069-01-15  0069-04-16
#>  7 Vespasian 0069-12-21  0079-06-24
#>  8 Titus     0079-06-24  0081-09-13
#>  9 Domitian  0081-09-14  0096-09-18
#> 10 Nerva     0096-09-18  0098-01-27
#> # … with 31 more rows
consolidate(emperors, "every", "every", resolve = "random", key = "ID")
#> # A tibble: 41 × 3
#>    ID        Beg        End       
#>    <chr>     <chr>      <chr>     
#>  1 Augustus  -XX31      0014-08-06
#>  2 Tiberius  0014-09-18 0037-02-26
#>  3 Claudius  0041-10-31 0054-02-12
#>  4 Nero      0054-11-27 0068-10-15
#>  5 Galba     0068-03-02 0069-02-25
#>  6 Otho      0069-01-15 0069-04-16
#>  7 Vespasian 0069-12-09 0079-07-04
#>  8 Titus     0079-06-24 0081-12-29
#>  9 Domitian  0081-05-11 0096-09-18
#> 10 Nerva     0096-09-18 0098-07-31
#> # … with 31 more rows
consolidate(emperors, "every", "every", resolve = c(Beg = "min", End = "max"), key = "ID")
#> # A tibble: 41 × 3
#>    ID        Beg         End       
#>    <chr>     <chr>       <chr>     
#>  1 Augustus  -XX26-01-16 0014-12-31
#>  2 Tiberius  -XX14       0037-12-31
#>  3 Claudius  0041-01-01  0054-12-31
#>  4 Nero      0054-01-01  0068-12-31
#>  5 Galba     0068-01-01  0069-12-31
#>  6 Otho      0069-01-01  0069-12-31
#>  7 Vespasian 0069-01-01  0079-12-31
#>  8 Titus     0079-01-01  0081-12-31
#>  9 Domitian  0081-01-01  0096-12-31
#> 10 Nerva     0096-01-01  0098-12-31
#> # … with 31 more rows
# }