Consolidate database into a single dataset

This function consolidates a set of datasets in a 'many* package' database into a single dataset with some combination of the rows, columns, and observations of the datasets in the database. The function includes separate arguments for the rows and columns, as well as for how to resolve conflicts for observations across datasets. This provides users with considerable flexibility in how they combine data. For example, users may wish to stick to units that appear in every dataset but include variables coded in any dataset, or units that appear in any dataset but only those variables that appear in every dataset. Even then there may be conflicts, as the actual unit-variable observations may differ from dataset to dataset. We offer a number of resolve methods that enable users to choose how conflicts between observations are resolved.

Usage

consolidate(
  database,
  rows = "any",
  cols = "any",
  resolve = "coalesce",
  key = "manyID"
)

Arguments

database

A database object from one of the many packages

rows

Which rows or units to retain. By default "any" (or all) units are retained, but another option is "every", which retains only those units that appear in all parent datasets.

cols

Which columns or variables to retain. By default "any" (or all) variables are retained, but another option is "every", which retains only those variables that appear in all parent datasets.

resolve

How should conflicts between observations be resolved? By default "coalesce", but other options include: "min", "max", "mean", "median", and "random". "coalesce" takes the first non-NA value. "max" takes the largest value. "min" takes the smallest value. "mean" takes the average value. "median" takes the median value. "random" takes a random value. For different variables to be resolved differently, you can specify the variables' names alongside how each is to be resolved in a list (e.g. resolve = c(var1 = "min", var2 = "max")). In this case, only the variables named will be resolved and returned.

key

An ID column to collapse by. By default "many_ID". Users can also specify multiple key variables in a list. For multiple key variables, the key variables must be present in all the datasets in the database (e.g. key = c("key1", "key2")). For equivalent key columns with different names across datasets, matching is possible if keys are declared (e.g. key = c("key1" = "key2")).

Value

A single tibble/data frame.

Examples

# \donttest{
consolidate(database = emperors, key = "ID")
#> # A tibble: 137 × 15
#>    ID        CityBirth  ProvinceBirth Rise  Cause Killer Era   Notes Verif Birth
#>    <chr>     <chr>      <chr>         <chr> <chr> <chr>  <chr> <chr> <chr> <chr>
#>  1 Augustus  Rome       Italia        Birt… Assa… Wife   Prin… birt… Redd… 0062…
#>  2 Tiberius  Rome       Italia        Birt… Assa… Other… Prin… birt… Redd… 0041…
#>  3 Caligula  Antitum    Italia        Birt… Assa… Senate Prin… assa… Redd… 0012…
#>  4 Claudius  Lugdunum   Gallia Lugdu… Birt… Assa… Wife   Prin… birt… Redd… 0009…
#>  5 Nero      Antitum    Italia        Birt… Suic… Senate Prin… NA    Redd… 0037…
#>  6 Galba     Terracina  Italia        Seiz… Assa… Other… Prin… birt… Redd… 0002…
#>  7 Otho      Terentinum Italia        Appo… Suic… Other… Prin… NA    NA    0032…
#>  8 Vitellius Rome       Italia        Seiz… Assa… Other… Prin… NA    NA    0015…
#>  9 Vespasian Falacrine  Italia        Seiz… Natu… Disea… Prin… NA    NA    0009…
#> 10 Titus     Rome       Italia        Birt… Natu… Disea… Prin… NA    NA    0039…
#> # … with 127 more rows, and 5 more variables: Death <chr>, FullName <chr>,
#> #   Dynasty <chr>, Beg <mdate>, End <mdate>
consolidate(database = favour(emperors, "UNRV"), rows = "every",
cols = "every", resolve = "coalesce", key = "ID")
#> # A tibble: 41 × 3
#>    ID        Beg     End    
#>    <chr>     <mdate> <mdate>
#>  1 Augustus  -0027   -0014  
#>  2 Tiberius  -0014   0037   
#>  3 Claudius  0041    0054   
#>  4 Nero      0054    0068   
#>  5 Galba     0068    0069   
#>  6 Otho      0069    0069   
#>  7 Vespasian 0069    0079   
#>  8 Titus     0079    0081   
#>  9 Domitian  0081    0096   
#> 10 Nerva     0096    0098   
#> # … with 31 more rows
consolidate(database = emperors, rows = "any", cols = "every",
resolve = "min", key = "ID")
#> # A tibble: 137 × 3
#>    ID        Beg         End       
#>    <chr>     <chr>       <chr>     
#>  1 Augustus  -0026-01-16 -0014     
#>  2 Tiberius  -0014       0037      
#>  3 Caligula  0037-03-18  0041-01-24
#>  4 Claudius  0041        0054      
#>  5 Nero      0054        0068      
#>  6 Galba     0068        0069      
#>  7 Otho      0069        0069      
#>  8 Vitellius 0069-04-17  0069-12-20
#>  9 Vespasian 0069        0079      
#> 10 Titus     0079        0081      
#> # … with 127 more rows
consolidate(database = emperors, rows = "every", cols = "any",
resolve = "max", key = "ID")
#> # A tibble: 41 × 15
#>    ID        CityBirth  ProvinceBirth Rise  Cause Killer Era   Notes Verif Birth
#>    <chr>     <chr>      <chr>         <chr> <chr> <chr>  <chr> <chr> <chr> <chr>
#>  1 Augustus  Rome       Italia        Birt… Assa… Wife   Prin… birt… Redd… 63 BC
#>  2 Tiberius  Rome       Italia        Birt… Assa… Other… Prin… birt… Redd… 42 BC
#>  3 Claudius  Lugdunum   Gallia Lugdu… Birt… Assa… Wife   Prin… birt… Redd… 10 BC
#>  4 Nero      Antitum    Italia        Birt… Suic… Senate Prin… NA    Redd… 37   
#>  5 Galba     Terracina  Italia        Seiz… Assa… Other… Prin… birt… Redd… 3 BC 
#>  6 Otho      Terentinum Italia        Appo… Suic… Other… Prin… NA    NA    32   
#>  7 Vespasian Falacrine  Italia        Seiz… Natu… Disea… Prin… NA    NA    9    
#>  8 Titus     Rome       Italia        Birt… Natu… Disea… Prin… NA    NA    39   
#>  9 Domitian  Rome       Italia        Birt… Assa… Court… Prin… NA    NA    51   
#> 10 Nerva     Narni      Italia        Appo… Natu… Disea… Prin… NA    NA    30   
#> # … with 31 more rows, and 5 more variables: Death <chr>, FullName <chr>,
#> #   Dynasty <chr>, Beg <chr>, End <chr>
consolidate(database = emperors, rows = "every", cols = "every",
resolve = "median", key = "ID")
#> # A tibble: 41 × 3
#>    ID        Beg     End         
#>    <chr>     <chr>   <chr>       
#>  1 Augustus  -0027   "0014\u001a"
#>  2 Tiberius  0014    "0037"      
#>  3 Claudius  0041    "0054"      
#>  4 Nero      0054    "0068"      
#>  5 Galba     0068    "0069"      
#>  6 Otho      0069-01 "0069-04"   
#>  7 Vespasian 0069    "0079"      
#>  8 Titus     0079    "0081"      
#>  9 Domitian  0081    "0096"      
#> 10 Nerva     0096    "0098"      
#> # … with 31 more rows
consolidate(database = emperors, rows = "every", cols = "every",
resolve = "mean", key = "ID")
#> # A tibble: 41 × 3
#>    ID        Beg         End       
#>    <chr>     <chr>       <chr>     
#>  1 Augustus  -0026-01-16 0014-08-19
#>  2 Tiberius  0014-09-18  0037-03-16
#>  3 Claudius  0041-01-25  0054-10-13
#>  4 Nero      0054-10-13  0068-06-09
#>  5 Galba     0068-06-08  0069-01-15
#>  6 Otho      0069-01-15  0069-04-16
#>  7 Vespasian 0069-12-21  0079-06-24
#>  8 Titus     0079-06-24  0081-09-13
#>  9 Domitian  0081-09-14  0096-09-18
#> 10 Nerva     0096-09-18  0098-01-27
#> # … with 31 more rows
consolidate(database = emperors, rows = "every", cols = "every",
resolve = "random", key = "ID")
#> # A tibble: 41 × 3
#>    ID        Beg           End       
#>    <chr>     <chr>         <chr>     
#>  1 Augustus  "-0031\u001a" -0014     
#>  2 Tiberius  "0014"        0037      
#>  3 Claudius  "0041"        0054      
#>  4 Nero      "0054"        0068-06-09
#>  5 Galba     "0068"        0069-01-15
#>  6 Otho      "0069-01-15"  0069-04-16
#>  7 Vespasian "0069-12-21"  0079-06-24
#>  8 Titus     "0079"        0081      
#>  9 Domitian  "0081"        0096      
#> 10 Nerva     "0096-09-18"  0098      
#> # … with 31 more rows
consolidate(database = emperors, rows = "every", cols = "every",
resolve = c(Beg = "min", End = "max"), key = "ID")
#> # A tibble: 41 × 3
#>    ID        Beg         End       
#>    <chr>     <chr>       <chr>     
#>  1 Augustus  -0026-01-16 0014-08-19
#>  2 Tiberius  -0014       0037-03-16
#>  3 Claudius  0041        0054-10-13
#>  4 Nero      0054        0068-06-09
#>  5 Galba     0068        0069-01-15
#>  6 Otho      0069        0069-04-16
#>  7 Vespasian 0069        0079-06-24
#>  8 Titus     0079        0081-09-13
#>  9 Domitian  0081        0096-09-18
#> 10 Nerva     0096        0098-01-27
#> # … with 31 more rows
consolidate(database = emperors, rows = "any", cols = "any",
resolve = c(Death = "max", Cause = "coalesce"),
key = c("ID", "Beg"))
#> # A tibble: 201 × 4
#>    ID        Beg         Cause          Death     
#>    <chr>     <mdate>     <chr>          <chr>     
#>  1 Augustus  -0026-01-16 Assassination  0014-08-19
#>  2 Tiberius  0014-09-18  Assassination  0037-03-16
#>  3 Caligula  0037-03-18  Assassination  0041-01-24
#>  4 Claudius  0041-01-25  Assassination  0054-10-13
#>  5 Nero      0054-10-13  Suicide        0068-06-09
#>  6 Galba     0068-06-08  Assassination  0069-01-15
#>  7 Otho      0069-01-15  Suicide        0069-04-16
#>  8 Vitellius 0069-04-17  Assassination  0069-12-20
#>  9 Vespasian 0069-12-21  Natural Causes 0079-06-24
#> 10 Titus     0079-06-24  Natural Causes 0081-09-13
#> # … with 191 more rows
# }