Drop only columns used in formula

A function between dplyr's transmute and mutate

Usage

transmutate(.data, ...)

Source

https://stackoverflow.com/questions/51428156/dplyr-mutate-transmute-drop-only-the-columns-used-in-the-formula

Arguments

.data

Data frame to pass to the function

...

Variables to pass to the function

Value

Data frame with mutated variables and none of the variables used in the mutations, but, unlike dplyr::transmute(), all other unnamed variables.

Examples

# \donttest{
pluck(emperors, "wikipedia")
#> # A tibble: 68 × 15
#>    ID       Begin End   FullName Birth Death CityBirth ProvinceBirth Rise  Cause
#>    <chr>    <mda> <mda> <chr>    <chr> <chr> <chr>     <chr>         <chr> <chr>
#>  1 Augustus -26-… 14-0… IMPERAT… 62-0… 14-0… Rome      Italia        Birt… Assa…
#>  2 Tiberius 14-0… 37-0… TIBERIV… 41-1… 37-0… Rome      Italia        Birt… Assa…
#>  3 Caligula 37-0… 41-0… GAIVS I… 12-0… 41-0… Antitum   Italia        Birt… Assa…
#>  4 Claudius 41-0… 54-1… TIBERIV… 9-08… 54-1… Lugdunum  Gallia Lugdu… Birt… Assa…
#>  5 Nero     54-1… 68-0… NERO CL… 37-1… 68-0… Antitum   Italia        Birt… Suic…
#>  6 Galba    68-0… 69-0… SERVIVS… 2-12… 69-0… Terracina Italia        Seiz… Assa…
#>  7 Otho     69-0… 69-0… MARCVS … 32-0… 69-0… Terentin… Italia        Appo… Suic…
#>  8 Vitelli… 69-0… 69-1… AVLVS V… 15-0… 69-1… Rome      Italia        Seiz… Assa…
#>  9 Vespasi… 69-1… 79-0… TITVS F… 9-11… 79-0… Falacrine Italia        Seiz… Natu…
#> 10 Titus    79-0… 81-0… TITVS F… 39-1… 81-0… Rome      Italia        Birt… Natu…
#> # ℹ 58 more rows
#> # ℹ 5 more variables: Killer <chr>, Dynasty <chr>, Era <chr>, Notes <chr>,
#> #   Verif <chr>
transmutate(emperors$wikipedia, Beginning = Begin)
#> # A tibble: 68 × 15
#>    ID      End   FullName Birth Death CityBirth ProvinceBirth Rise  Cause Killer
#>    <chr>   <mda> <chr>    <chr> <chr> <chr>     <chr>         <chr> <chr> <chr> 
#>  1 August… 14-0… IMPERAT… 62-0… 14-0… Rome      Italia        Birt… Assa… Wife  
#>  2 Tiberi… 37-0… TIBERIV… 41-1… 37-0… Rome      Italia        Birt… Assa… Other…
#>  3 Caligu… 41-0… GAIVS I… 12-0… 41-0… Antitum   Italia        Birt… Assa… Senate
#>  4 Claudi… 54-1… TIBERIV… 9-08… 54-1… Lugdunum  Gallia Lugdu… Birt… Assa… Wife  
#>  5 Nero    68-0… NERO CL… 37-1… 68-0… Antitum   Italia        Birt… Suic… Senate
#>  6 Galba   69-0… SERVIVS… 2-12… 69-0… Terracina Italia        Seiz… Assa… Other…
#>  7 Otho    69-0… MARCVS … 32-0… 69-0… Terentin… Italia        Appo… Suic… Other…
#>  8 Vitell… 69-1… AVLVS V… 15-0… 69-1… Rome      Italia        Seiz… Assa… Other…
#>  9 Vespas… 79-0… TITVS F… 9-11… 79-0… Falacrine Italia        Seiz… Natu… Disea…
#> 10 Titus   81-0… TITVS F… 39-1… 81-0… Rome      Italia        Birt… Natu… Disea…
#> # ℹ 58 more rows
#> # ℹ 5 more variables: Dynasty <chr>, Era <chr>, Notes <chr>, Verif <chr>,
#> #   Beginning <mdate>
# }